Fast Cox Proportional Hazards Regression Model

Fits a Cox proportional hazards regression model. An extension to use (or not) your GPU to speed up calculations, in particular for bootstrap.

Use summary() method to see confidence interval for covariates with two process : normal distribution and bootstrap (if bootstrap > 1).

coxphGPU(formula, data, ties = c("efron", "breslow"), bootstrap = 1,
         batchsize = 0, init, all.results = FALSE, control,
         singular.ok = TRUE, model = FALSE, x = FALSE, y = TRUE, ...)

# S3 method for coxphGPU
summary(object, ..., conf.int = 0.95, scale = 1)

Arguments

formula

a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv function.

data

a data.frame in which to interpret the variables named in the formula, or in the subset and the weights argument.

ties

a character string specifying the method for tie handling. If there are no tied death times all the methods are equivalent. Nearly all Cox regression programs use the Breslow method by default, but not this one. The Efron approximation is used as the default here, it is more accurate when dealing with tied death times, and is as efficient computationally. The ``exact partial likelihood'' is equivalent to a conditional logistic model, and is appropriate when the times are a small set of discrete values. See further below.

bootstrap

Number of repeats for the bootstrap cross-validation.

batchsize

Number of bootstrap copies that should be handled at a time. Defaults to 0, which means that we handle all copies at once. If you run into out of memory errors, please consider using batchsize=100, 10 or 1.

init

vector of initial values of the iteration. Default initial value is zero for all variables.

all.results

Post-processing calculations. If TRUE, coxphGPU returns linears.predictors, wald.test, concordance for all bootstraps. Default to FALSE if bootstraps.

control

Object of class coxph.control specifying iteration limit and other control options. Default is coxph.control(...).

singular.ok

logical value indicating how to handle collinearity in the model matrix. If TRUE, the program will automatically skip over columns of the X matrix that are linear combinations of earlier columns. In this case the coefficients for such columns will be NA, and the variance matrix will contain zeros. For ancillary calculations, such as the linear predictor, the missing coefficients are treated as zeros.

model

logical value: if TRUE, the model frame is returned in component model.

x

logical value: if TRUE, the x matrix is returned in component x.

y

logical value: if TRUE, the response vector is returned in component y.

...

Other arguments for methods.

object

a coxphGPU object

conf.int

level for computation of the confidence intervals. If set to FALSE no confidence intervals are printed

scale

vector of scale factors for the coefficients, defaults to 1. The printed coefficients, se, and confidence intervals will be associated with one scale unit.

Value

A coxphGPU object representing the fit.

With summary() :

conf.int: a matrix with one row for each coefficient, containing the confidence limits for exp(coef).

conf.int_bootstrap: confidence limits for exp(coef) determined by bootstrap

logtest, sctest, waldtest: the overall likelihood ratio, score, and Wald test statistics for the model

concordance: the concordance statistic and its standard error

rsq: an approximate R^2 based on Nagelkirke (Biometrika 1991).

References

Therneau T (2021). A Package for Survival Analysis in R. R package version 3.2-13

See also

Examples

if (FALSE) {
library(survival)
library(WCE)
data(drugdata)

## Check CUDA drivers (if FALSE you use CPU)
use_cuda()

## Cox Proportional Hazards without bootstrap
coxphGPU(Surv(Start, Stop, Event) ~ sex + age,
         data = drugdata,
         bootstrap = 1)

## Cox Proportional Hazards with bootstrap

if (use_cuda()) {
  n_bootstrap <- 1000
  batchsize <- 200
} else {
  n_bootstrap <- 50
  batchsize <- 10
}

coxph_bootstrap <- coxphGPU(Surv(Start, Stop, Event) ~ sex + age,
                            data = drugdata,
                            bootstrap = n_bootstrap,
                            batchsize = batchsize)

summary(coxph_bootstrap)
}