Predict method for coxphGPU

Compute fitted values and regression terms for a model fitted by coxphGPU

# S3 method for coxphGPU
predict(
  object,
  newdata,
  type = c("lp", "risk", "expected", "terms", "survival"),
  se.fit = FALSE,
  na.action = na.pass,
  terms = names(object$assign),
  collapse,
  reference = c("strata", "sample", "zero"),
  ...
)

Arguments

object

the results of a coxphGPU fit.

newdata

Optional new data at which to do predictions. If absent predictions are for the data frame used in the original fit. When coxph has been called with a formula argument created in another context, i.e., coxph has been called within another function and the formula was passed as an argument to that function, there can be problems finding the data set. See the note below.

type

the type of predicted value. Choices are the linear predictor ("lp"), the risk score exp(lp) ("risk"), the expected number of events given the covariates and follow-up time ("expected"), and the terms of the linear predictor ("terms"). The survival probability for a subject is equal to exp(-expected).

se.fit

if TRUE, pointwise standard errors are produced for the predictions.

na.action

applies only when the newdata argument is present, and defines the missing value action for the new data. The default is to include all observations. When there is no newdata, then the behavior of missing is dictated by the na.action option of the original fit.

terms

if type="terms", this argument can be used to specify which terms should be included; the default is all.

collapse

optional vector of subject identifiers. If specified, the output will contain one entry per subject rather than one entry per observation.

reference

reference for centering predictions, see details below

...

For future methods

See also

Examples

if (FALSE) {
library(survival)
options(na.action = na.exclude) # retain NA in predictions
fit <- coxphGPU(Surv(time, status) ~ age + ph.ecog + strata(inst), lung)
predict(fit, type = "lp")
predict(fit, type = "expected")
predict(fit, type = "risk", se.fit = TRUE)
predict(fit, type = "terms", se.fit = TRUE)
}