Fast WCE

New implementation of the Weighted Cumulative Exposure model (see @details), compatible with GPU to accelerate calculation speed and work with large datasets.

Use summary() and plot() methods to see results and risk function.

wceGPU(data, nknots, cutoff, constrained = FALSE, aic = FALSE, id,
       event, start, stop, expos, covariates = NULL, nbootstraps = 1,
       batchsize = 0, confint = 0.95, controls = NULL, ...)

# S3 method for wceGPU
summary(object, ...)

# S3 method for wceGPU
plot(x, ..., hist.covariates = FALSE)

# S3 method for wceGPU
confint(object, parm, level = 0.95, ..., digits = 3)

Arguments

data

A data frame in an interval (long) format, in which each line corresponds to one and only one time unit for a given individual.

nknots

Corresponds to the number(s) of interior knots for the cubic splines to estimate the weight function. For example, if nknots is set to 2, then a model with two interior knots is fitted.

cutoff

Integer. Time window over which the WCE model is estimated. Corresponds to the length of the estimated weight function.

constrained

Controls whether the weight function should be constrained to smoothly go to zero. Set to FALSE for unconstrained models, to 'Right' or 'R' to constrain the weight function to smoothly go to zero for exposure remote in time, and to 'Left' or 'L' to constrain the weight function to start a zero for the current values.

aic

Logical. If TRUE, then the AIC is used to select the best fitting model among those estimated for the different numbers of interior knots requested with nknots. If FALSE, then the BIC is used instead of the AIC. Default to FALSE (BIC). Note that the BIC implemented in WCE is the version suggested by Volinsky and Raftery in Biometrics (2000), which corresponds to BIC = 2 * log(PL) + p * log(d) where PL is the model's partial likelihood, p is the number of estimated parameters and d is the number of uncensored events. See Sylvestre and Abrahamowicz (2009) for more details.

id

Name of the variable in data corresponding to the identification of subjects.

event

Name of the variable in data corresponding to event indicator. Must be coded 1 = event and 0 = no event.

start

Name of the variable in data corresponding to the starting time for the interval. Corresponds to time argument in function Surv in the survival package.

stop

Name of the variable in data corresponding to the ending time for the interval. Corresponds to time2 argument in function Surv in the survival package.

expos

Name of the variable in data corresponding to the exposure variable.

covariates

Optional. Vector of characters corresponding to the name(s) of the variable(s) in data corresponding to the covariate(s) to be included in the model. Default to NULL, which corresponds to fitting model(s) without covariates.

nbootstraps

Number of repeats for the bootstrap cross-validation.

batchsize

Number of bootstrap copies that should be handled at a time. Defaults to 0, which means that we handle all copies at once. If you run into out of memory errors, please consider using batchsize=100, 10 or 1.

confint

Level for confidence intervals. Default to 0.95.

controls

List corresponding to the control parameters to be passed to the coxph function. See coxph.control for more details.

...

additional argument(s) for methods.

object

wceGPU object.

x

wceGPU object.

hist.covariates

show histogram for each covariates if you use bootstrap.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

digits

significant digits to print.

Value

WCE results

Details

WCE implements a flexible method for modeling cumulative effects of time-varying exposures, weighted according to their relative proximity in time, and represented by time-dependent covariates. The current implementation estimates the weight function in the Cox proportional hazards model. The function that assigns weights to doses taken in the past is estimated using cubic regression splines.

References

Sylvestre MP, Abrahamowicz M. Flexible modeling of the cumulative effects of time-dependent exposures on the hazard. Stat Med. 2009 Nov 30;28(27):3437-53.

Examples

if (FALSE) {
# Dataset
drugdata <- WCE::drugdata

# WCE model
wce_gpu <- wceGPU(data = drugdata, nknots = 1, cutoff = 90, id = "Id",
                  event = "Event", start = "Start", stop = "Stop",
                  expos = "dose", covariates = c("age", "sex"),
                  constrained = FALSE, aic = FALSE, confint = 0.95,
                  nbootstraps = 1, batchsize = 0)

# Results
wce_gpu
summary(wce_gpu)

# See estimated weight function
plot(wce_gpu)

# WCE model with bootstrap (example with 20 bootstraps, but normally
# nbootstraps > 500)
wce_gpu_bootstrap <- wceGPU(data = drugdata, nknots = 1, cutoff = 90,
                            id = "Id", event = "Event", start = "Start",
                            stop = "Stop", expos = "dose",
                            covariates = c("age", "sex"),
                            constrained = FALSE, aic = FALSE, confint = 0.95,
                            nbootstraps = 20, batchsize = 0)

# See confidence bands for the estimated weight function due to bootstrap
plot(wce_gpu_bootstrap)

# All estimated coefficients in bootstrap
coef(wce_gpu_bootstrap)

# Estimate a HR (Exposed at a dose vs. unexposed)
exposed   <- rep(1, 90)
unexposed <- rep(0, 90)

HR(wce_gpu_bootstrap, exposed, unexposed)
}